Separate Monolith public endpoints
Use the `/public` prefix for the Monolith views that need to be public.
